+++ FreshservicePS/Public/Get-FreshServiceChange.ps1
@@ -458,10 +458,10 @@ function Get-FreshServiceChange {
 
                 $result = Invoke-FreshworksRestMethod @params
 
-                $content = $result.Content |
-                                ConvertFrom-Json
+                if ($result.Content) {
+                    $content = $result.Content |
+                                    ConvertFrom-Json
 
-                if ($content) {
                     #API returns singluar or plural property based on the number of records, parse to get property returned.
                     #When using Filter, the API also returns a Total property, so we are filtering here to only return change or changes property
                     $objProperty = $content[0].PSObject.Properties |