+++ FreshservicePS/Public/Get-FreshServiceCustomObject.ps1
@@ -118,10 +118,10 @@ function Get-FreshServiceCustomObject {
 
                 $result = Invoke-FreshworksRestMethod @params
 
-                $content = $result.Content |
-                                ConvertFrom-Json
+                if ($result.Content) {
+                    $content = $result.Content |
+                                    ConvertFrom-Json
 
-                if ($content) {
                     #API returns singluar or plural property based on the number of records, parse to get property returned.
                     $objProperty = $content[0].PSObject.Properties.Name |
                                         Where-Object -FilterScript {$_ -like 'custom_object*'}