+++ FreshservicePS/Public/Get-FreshServiceCustomObjectRecord.ps1
@@ -119,10 +119,10 @@ function Get-FreshServiceCustomObjectRecord {
 
                 $result = Invoke-FreshworksRestMethod @params
 
-                $content = $result.Content |
-                                ConvertFrom-Json
+                if ($result.Content) {
+                    $content = $result.Content |
+                                    ConvertFrom-Json
 
-                if ($content) {
                     #API returns singluar or plural property based on the number of records, parse to get property returned.
                     $objProperty = $content[0].PSObject.Properties.Name  |
                                         Where-Object -FilterScript {$_ -like 'record*'}