+++ FreshservicePS/Public/Get-FreshServiceInfo.ps1
@@ -31,48 +31,22 @@ function Get-FreshServiceInfo {
 
         $qry = [System.Web.HttpUtility]::ParseQueryString([String]::Empty)
         $uri = [System.UriBuilder]('{0}/business_hours' -f $PrivateData['FreshserviceBaseUri'])
-        # $enablePagination = $true
 
     }
     process {
 
         try {
 
-            # if ($enablePagination) {
-            #     $qry['page'] = $page
-            #     $qry['per_page'] = $per_page
-            # }
-
             $uri.Query = $qry.ToString()
-
             $uri = $uri.Uri.AbsoluteUri
 
-            # $results = do {
+            $params = @{
+                Uri         = $uri
+                Method      = 'GET'
+                ErrorAction = 'Stop'
+            }
 
-                $params = @{
-                    Uri         = $uri
-                    Method      = 'GET'
-                    ErrorAction = 'Stop'
-                }
-
-                $result = Invoke-FreshworksRestMethod @params
-
-                # $content = $result.Content |
-                #                 ConvertFrom-Json
-
-                # if ($content) {
-                #     #API returns singluar or plural property based on the number of records, parse to get property returned.
-                #     $objProperty = $content[0].PSObject.Properties.Name
-                #     Write-Verbose -Message ("Returning {0} property with count {1}" -f $objProperty, $content."$($objProperty)".Count)
-                #     $content."$($objProperty)"
-                # }
-
-                # if ($result.Headers.Link) {
-                #     $uri = [regex]::Matches($result.Headers.Link,'<(?<Uri>.*)>')[0].Groups['Uri'].Value
-                # }
-
-            # }
-            # until (!$result.Headers.Link)
+            $result = Invoke-FreshworksRestMethod @params
 
         }
         catch {