+++ FreshservicePS/Public/New-FreshServiceNote.ps1
@@ -79,7 +79,7 @@ function New-FreshServiceNote {
     }
     process {
 
-        $uri = [System.UriBuilder]('{0}/{1}s/{2}/notes' -f $PrivateData['FreshserviceBaseUri'], $Type.ToLower(),$parent_id)
+        $uri = [System.UriBuilder]('{0}/{1}s/{2}\notes' -f $PrivateData['FreshserviceBaseUri'], $Type.ToLower(),$parent_id)
 
         [void]$PSBoundParameters.Remove('type')
         [void]$PSBoundParameters.Remove('parent_id')
@@ -108,10 +108,10 @@ function New-FreshServiceNote {
 
                 $result = Invoke-FreshworksRestMethod @params
 
-                $content = $result.Content |
-                                ConvertFrom-Json
+                if ($result.Content) {
+                    $content = $result.Content |
+                                    ConvertFrom-Json
 
-                if ($content) {
                     #API returns singluar or plural property based on the number of records, parse to get property returned.
                     $objProperty = $content[0].PSObject.Properties.Name
                     Write-Verbose -Message ("Returning {0} property with count {1}" -f $objProperty, $content."$($objProperty)".Count)