+++ FreshservicePS/Public/Set-FreshServiceNote.ps1
@@ -87,7 +87,7 @@ function Set-FreshServiceNote {
     }
     process {
 
-        $uri = [System.UriBuilder]('{0}/{1}s/{2}/notes/{3}' -f $PrivateData['FreshserviceBaseUri'], $type.ToLower(),$parent_id,$id)
+        $uri = [System.UriBuilder]('{0}/{1}s/{2}\notes/{3}' -f $PrivateData['FreshserviceBaseUri'], $type.ToLower(),$parent_id,$id)
 
         [void]$PSBoundParameters.Remove('type')
         [void]$PSBoundParameters.Remove('parent_id')
@@ -112,10 +112,10 @@ function Set-FreshServiceNote {
 
                 $result = Invoke-FreshworksRestMethod @params
 
-                $content = $result.Content |
-                                ConvertFrom-Json
+                if ($result.Content) {
+                    $content = $result.Content |
+                                    ConvertFrom-Json
 
-                if ($content) {
                     #API returns singluar or plural property based on the number of records, parse to get property returned.
                     $objProperty = $content[0].PSObject.Properties.Name
                     Write-Verbose -Message ("Returning {0} property with count {1}" -f $objProperty, $content."$($objProperty)".Count)